requirements.txt
Adafruit-Blinka==8.58.1
adafruit-circuitpython-busdevice==5.2.12
adafruit-circuitpython-connectionmanager==3.1.4
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-thermal-printer==1.4.0
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.79.0
Adafruit-PureIO==1.1.11
appdirs==1.4.4
argcomplete==3.6.2
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorzero==2.0
contourpy==1.3.2
cycler==0.12.1
dependency-injector==4.46.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.58.0
fsspec==2025.3.2
gpiozero==2.0.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
ncnn==1.0.20250503
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.1.0
portalocker==3.1.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyftdi==0.56.0
pyparsing==3.2.3
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-escpos==3.1
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.2
requests==2.32.3
RPi.GPIO==0.7.1
scipy==1.15.3
simpleaudio==1.0.4
six==1.17.0
sympy==1.14.0
sysv-ipc==1.1.0
tabulate==0.9.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.136
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
